+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Full changelog for Bootstrap 4 changelog
 
+## v0.02 - Badge, Button and Progress bar - 2017-11-29
+
+* Added Badge view helper.
+* Added Button view helper.
+* Added Progress bar view helper.
+* Updated README, examples for each view helper and callable methods. 
+
+## v0.01 - Jumbotron - 2017-11-13
+
+* Dev release for testing, added Jumbotron view helper.
+
 ## Development
 
 * Repo setup, git  files, initial README, CHANGELOG, LICENSE and library requirements.